fir_p.c: Substantial rewrite to improve realtime performance.
Now utilizes multiple threads along with a different partitioning
scheme to achieve much lower realtime latency.

Additionally, frequency-domain delay lines and frequency-domain
summing significantly improve throughput (even when single-threaded)
by reducing the number of required FFT/IFFT operations.

+++ fir_p.h
@@ -22,7 +22,7 @@
 #include "dsp.h"
 #include "effect.h"
 
-struct effect * fir_p_effect_init_with_filter(const struct effect_info *, const struct stream_info *, const char *, sample_t *, int, ssize_t, ssize_t);
+struct effect * fir_p_effect_init_with_filter(const struct effect_info *, const struct stream_info *, const char *, sample_t *, int, ssize_t, int);
 struct effect * fir_p_effect_init(const struct effect_info *, const struct stream_info *, const char *, const char *, int, const char *const *);
 
 #endif
